refactor(earn): move vaults components to shared
