+++ features/earn/shared/types.tsx
