+++ features/earn/shared/vault-description/index.ts
