+++ features/earn/shared/vault-description/styles.tsx
