+++ features/earn/shared/vault-description/vault-description.tsx
