+++ features/earn/shared/vault-header/index.ts
