+++ features/earn/shared/vault-header/styles.tsx
