+++ features/earn/shared/vault-partners/styles.tsx
