+++ features/earn/shared/vault-stats/styles.tsx
