+++ features/earn/shared/vault-stats/vault-stats.tsx
